pyproject.toml
[project]
name = "canismajor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pyyaml>=6.0.2",
    "aiofiles>=24.1.0",
    "aioconsole>=0.8.1",
]

[project.optional-dependencies]
extra = ["qrcode>=8.1"]
sra = ["speechrecognition>=3.14.1", "sounddevice>=0.5.1", "jellyfish>=1.1.3"]
